+++ independentDraw/drawBoxWithRoundedCornersIntoThePathWithDensity.go
@@ -8,24 +8,24 @@ import (
 // en: This function create a path of a box with rounded corners into the platform using density
 //
 // pt_br: Esta função cria o caminho de um quadrado com cantos arredondados na plataforma usando densidade
-func DrawBoxWithRoundedCornersIntoThePathWithDensity(platform iotmaker_platform_IDraw.IDraw, x, y, width, height, border int, density interface{}, iDensity iotmaker_platform_coordinate.IDensity) {
+func DrawBoxWithRoundedCornersIntoThePathWithDensity(platform iotmaker_platform_IDraw.IDraw, x, y, width, height, border float64, density interface{}, iDensity iotmaker_platform_coordinate.IDensity) {
 	densityCalc := iDensity
 	densityCalc.SetDensityFactor(density)
 
 	densityCalc.Set(x)
-	x = densityCalc.Int()
+	x = densityCalc.Float64()
 
 	densityCalc.Set(y)
-	y = densityCalc.Int()
+	y = densityCalc.Float64()
 
 	densityCalc.Set(width)
-	width = densityCalc.Int()
+	width = densityCalc.Float64()
 
 	densityCalc.Set(height)
-	height = densityCalc.Int()
+	height = densityCalc.Float64()
 
 	densityCalc.Set(border)
-	border = densityCalc.Int()
+	border = densityCalc.Float64()
 
 	DrawBoxWithRoundedCornersIntoThePath(platform, x, y, width, height, border)
 }
